+++ crates/openapi/src/routes/mandates.rs
@@ -66,6 +66,9 @@ pub async fn retrieve_mandates_list() {}
 #[utoipa::path(
     post,
     path = "/customers/{customer_id}/mandates",
+    params(
+        ("customer_id" = String, Path, description = "The unique identifier for the customer")
+    ),
     responses(
         (status = 200, description = "List of retrieved mandates for a customer", body = Vec<MandateResponse>),
         (status = 400, description = "Invalid Data"),