+++ crates/router/src/routes/lock_utils.rs
@@ -43,6 +43,7 @@ pub enum ApiIdentifier {
     Hypersense,
     PaymentMethodSession,
     ProcessTracker,
+    Proxy,
 }
 
 impl From<Flow> for ApiIdentifier {
@@ -340,6 +341,7 @@ impl From<Flow> for ApiIdentifier {
             | Flow::PaymentMethodSessionUpdate => Self::PaymentMethodSession,
 
             Flow::RevenueRecoveryRetrieve => Self::ProcessTracker,
+            Flow::Proxy => Self::Proxy,
         }
     }
 }