+++ migrations/2025-05-21-134155_add_three_ds_decision_rule_algorithm_to_business_profile/down.sql
@@ -1,0 +1,3 @@
+-- This file should undo anything in `up.sql`
+ALTER TABLE business_profile
+DROP COLUMN three_ds_decision_rule_algorithm;